Split harmburger menu from map activity, presenter , interactor
