+++ opensrp-reveal/src/main/java/org/smartregister/reveal/view/BaseMapActivity.java
@@ -1,14 +1,11 @@
-package org.smartregister.reveal.activity;
+package org.smartregister.reveal.view;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import org.smartregister.reveal.view.RevealMapView;
 
 /**
  * Created by samuelgithengi on 11/20/18.
  */
-public abstract class BaseMapActivity extends AppCompatActivity {
+public abstract class BaseMapActivity extends BaseDrawerActivity {
 
     protected RevealMapView kujakuMapView;
