+++ opensrp-reveal/src/main/java/org/smartregister/reveal/view/ListTasksActivity.java
@@ -1,24 +1,16 @@
 package org.smartregister.reveal.view;
 
-import android.app.ProgressDialog;
 import android.content.BroadcastReceiver;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.location.Location;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.StringRes;
-import android.support.constraint.ConstraintLayout;
-import android.support.design.widget.NavigationView;
 import android.support.design.widget.Snackbar;
 import android.support.v4.content.LocalBroadcastManager;
-import android.support.v4.util.Pair;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.CardView;
 import android.text.TextUtils;
@@ -26,10 +18,8 @@ import android.util.Log;
 import android.view.Gravity;
 import android.view.MotionEvent;
 import android.view.View;
-import android.view.ViewTreeObserver;
 import android.widget.FrameLayout;
 import android.widget.ImageButton;
-import android.widget.LinearLayout.LayoutParams;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -42,19 +32,15 @@ import com.mapbox.mapboxsdk.geometry.LatLng;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback;
 import com.mapbox.mapboxsdk.style.sources.GeoJsonSource;
-import com.vijay.jsonwizard.customviews.TreeViewDialog;
 
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.smartregister.domain.FetchStatus;
 import org.smartregister.receiver.SyncStatusBroadcastReceiver;
-import org.smartregister.repository.AllSharedPreferences;
-import org.smartregister.reveal.BuildConfig;
 import org.smartregister.reveal.R;
-import org.smartregister.reveal.activity.BaseMapActivity;
 import org.smartregister.reveal.activity.RevealJsonFormActivity;
 import org.smartregister.reveal.application.RevealApplication;
+import org.smartregister.reveal.contract.BaseDrawerContract;
 import org.smartregister.reveal.contract.ListTaskContract;
 import org.smartregister.reveal.contract.UserLocationContract.UserLocationView;
 import org.smartregister.reveal.model.CardDetails;
@@ -62,13 +48,6 @@ import org.smartregister.reveal.presenter.ListTaskPresenter;
 import org.smartregister.reveal.util.Constants.Action;
 import org.smartregister.reveal.util.Constants.TaskRegister;
 import org.smartregister.reveal.util.RevealMapHelper;
-import org.smartregister.util.Utils;
-
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-import java.util.Locale;
 
 import io.ona.kujaku.callbacks.OnLocationComponentInitializedCallback;
 import io.ona.kujaku.utils.Constants;
@@ -89,27 +68,15 @@ public class ListTasksActivity extends BaseMapActivity implements ListTaskContra
 
     private static final String TAG = "ListTasksActivity";
 
-    private AllSharedPreferences sharedPreferences;
-
     private ListTaskPresenter listTaskPresenter;
 
     private View rootView;
-
-    private TextView campaignTextView;
-    private TextView operationalAreaTextView;
-    private TextView districtTextView;
-    private TextView facilityTextView;
-    private TextView operatorTextView;
 
     private GeoJsonSource geoJsonSource;
 
     private GeoJsonSource selectedGeoJsonSource;
 
-    private ProgressDialog progressDialog;
-
     private MapboxMap mMapboxMap;
-
-    private DrawerLayout mDrawerLayout;
 
     private CardView structureInfoCardView;
     private TextView tvSprayStatus;
@@ -132,8 +99,6 @@ public class ListTasksActivity extends BaseMapActivity implements ListTaskContra
 
         listTaskPresenter = new ListTaskPresenter(this);
         rootView = findViewById(R.id.content_frame);
-
-        sharedPreferences = RevealApplication.getInstance().getContext().allSharedPreferences();
 
         initializeMapView(savedInstanceState);
         initializeDrawerLayout();
@@ -239,86 +204,6 @@ public class ListTasksActivity extends BaseMapActivity implements ListTaskContra
         }
     }
 
-    private void initializeDrawerLayout() {
-
-        mDrawerLayout = findViewById(R.id.drawer_layout);
-
-        ImageButton mDrawerMenuButton = findViewById(R.id.drawerMenu);
-        mDrawerMenuButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mDrawerLayout.openDrawer(GravityCompat.START);
-            }
-        });
-
-        mDrawerLayout.addDrawerListener(new DrawerLayout.DrawerListener() {
-            @Override
-            public void onDrawerSlide(@NonNull View drawerView, float slideOffset) {//do nothing
-            }
-
-            @Override
-            public void onDrawerOpened(@NonNull View drawerView) {//do nothing
-            }
-
-            @Override
-            public void onDrawerClosed(@NonNull View drawerView) {
-                listTaskPresenter.onDrawerClosed();
-            }
-
-            @Override
-            public void onDrawerStateChanged(int newState) {//do nothing
-            }
-        });
-
-        NavigationView navigationView = findViewById(R.id.nav_view);
-        View headerView = navigationView.getHeaderView(0);
-
-        headerView.getViewTreeObserver().addOnGlobalLayoutListener(new ViewTreeObserver.OnGlobalLayoutListener() {
-            @Override
-            public void onGlobalLayout() {
-                headerView.getViewTreeObserver().removeOnGlobalLayoutListener(this);
-                int minimumOperatorMargin = getResources().getDimensionPixelSize(R.dimen.operator_top_margin);
-                int screenHeightPixels = getResources().getDisplayMetrics().heightPixels;
-                //if content of hamburger menu is bigger than screen; scroll content
-                if (screenHeightPixels < headerView.getHeight() + minimumOperatorMargin) {
-                    headerView.setLayoutParams(new LayoutParams(LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT));
-                    View operator = headerView.findViewById(R.id.operator_label);
-                    ConstraintLayout.LayoutParams params = (ConstraintLayout.LayoutParams) operator.getLayoutParams();
-                    params.height = ConstraintLayout.LayoutParams.WRAP_CONTENT;
-                    operator.setLayoutParams(params);
-                } else {//content of hamburger menu fits on screen; set menu height to screen height
-                    headerView.setLayoutParams(new LayoutParams(LayoutParams.MATCH_PARENT,
-                            screenHeightPixels - getResources().getDimensionPixelSize(R.dimen.hamburger_margin)));
-                }
-            }
-        });
-
-        try {
-            ((TextView) headerView.findViewById(R.id.application_version))
-                    .setText(getString(R.string.app_version, Utils.getVersion(this)));
-        } catch (PackageManager.NameNotFoundException e) {
-            Log.e(TAG, e.getMessage(), e);
-        }
-
-        String buildDate = new SimpleDateFormat("dd/MM/yyyy", Locale.getDefault())
-                .format(new Date(BuildConfig.BUILD_TIMESTAMP));
-        ((TextView) headerView.findViewById(R.id.application_updated)).setText(getString(R.string.app_updated, buildDate));
-
-        campaignTextView = headerView.findViewById(R.id.campaign_selector);
-        operationalAreaTextView = headerView.findViewById(R.id.operational_area_selector);
-        districtTextView = headerView.findViewById(R.id.district_label);
-        facilityTextView = headerView.findViewById(R.id.facility_label);
-        operatorTextView = headerView.findViewById(R.id.operator_label);
-
-        operationalAreaTextView.setOnClickListener(this);
-
-        campaignTextView.setOnClickListener(this);
-
-        headerView.findViewById(R.id.logout_button).setOnClickListener(this);
-        headerView.findViewById(R.id.sync_button).setOnClickListener(this);
-
-    }
-
     @Override
     public void onClick(View v) {
         if (v.getId() == R.id.operational_area_selector)
@@ -329,7 +214,7 @@ public class ListTasksActivity extends BaseMapActivity implements ListTaskContra
             RevealApplication.getInstance().logoutCurrentUser();
         else if (v.getId() == R.id.sync_button) {
             org.smartregister.reveal.util.Utils.startImmediateSync();
-            mDrawerLayout.closeDrawer(GravityCompat.START);
+            closeDrawerLayout();
         } else if (v.getId() == R.id.btn_add_structure) {
             listTaskPresenter.onAddStructureClicked();
         } else if (v.getId() == R.id.change_spray_status) {
@@ -369,68 +254,12 @@ public class ListTasksActivity extends BaseMapActivity implements ListTaskContra
     }
 
     @Override
-    public void showOperationalAreaSelector(Pair<String, ArrayList<String>> locationHierarchy) {
-        try {
-            TreeViewDialog treeViewDialog = new TreeViewDialog(ListTasksActivity.this,
-                    R.style.AppTheme_WideDialog,
-                    new JSONArray(locationHierarchy.first), locationHierarchy.second, locationHierarchy.second);
-            treeViewDialog.setCancelable(true);
-            treeViewDialog.setCanceledOnTouchOutside(true);
-            treeViewDialog.setOnDismissListener(new DialogInterface.OnDismissListener() {
-                @Override
-                public void onDismiss(DialogInterface dialog) {
-                    listTaskPresenter.onOperationalAreaSelectorClicked(treeViewDialog.getName());
-                }
-            });
-            treeViewDialog.show();
-        } catch (JSONException e) {
-            Log.e(TAG, e.getMessage());
-        }
-
-    }
-
-    @Override
-    public void showCampaignSelector(List<String> campaigns, String entireTreeString) {
-        try {
-            TreeViewDialog treeViewDialog = new TreeViewDialog(ListTasksActivity.this,
-                    R.style.AppTheme_WideDialog,
-                    new JSONArray(entireTreeString), new ArrayList<>(campaigns), new ArrayList<>(campaigns));
-            treeViewDialog.show();
-            treeViewDialog.setCanceledOnTouchOutside(true);
-            treeViewDialog.setOnDismissListener(new DialogInterface.OnDismissListener() {
-                @Override
-                public void onDismiss(DialogInterface dialog) {
-                    listTaskPresenter.onCampaignSelectorClicked(treeViewDialog.getValue(), treeViewDialog.getName());
-                }
-            });
-            treeViewDialog.show();
-        } catch (JSONException e) {
-            Log.e(TAG, e.getMessage());
-        }
-    }
-
-    @Override
     public void setGeoJsonSource(@NonNull FeatureCollection featureCollection, Geometry operationalAreaGeometry) {
         if (geoJsonSource != null) {
             geoJsonSource.setGeoJson(featureCollection);
             if (operationalAreaGeometry != null) {
                 mMapboxMap.setCameraPosition(mMapboxMap.getCameraForGeometry(operationalAreaGeometry));
             }
-        }
-    }
-
-    @Override
-    public void lockNavigationDrawerForSelection() {
-        mDrawerLayout.openDrawer(GravityCompat.START);
-        mDrawerLayout.setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_OPEN);
-
-    }
-
-    @Override
-    public void unlockNavigationDrawer() {
-        if (mDrawerLayout.getDrawerLockMode(GravityCompat.START) == DrawerLayout.LOCK_MODE_LOCKED_OPEN) {
-            mDrawerLayout.closeDrawer(GravityCompat.START);
-            mDrawerLayout.setDrawerLockMode(DrawerLayout.LOCK_MODE_UNLOCKED);
         }
     }
 
@@ -525,53 +354,6 @@ public class ListTasksActivity extends BaseMapActivity implements ListTaskContra
         }
     }
 
-    @Override
-    public void setCampaign(String campaign) {
-        campaignTextView.setText(campaign);
-    }
-
-    @Override
-    public void setOperationalArea(String operationalArea) {
-        operationalAreaTextView.setText(operationalArea);
-    }
-
-    @Override
-    public void setDistrict(String district) {
-        org.smartregister.reveal.util.Utils.setTextViewText(districtTextView, R.string.district, district);
-    }
-
-    @Override
-    public void setFacility(String facility) {
-        org.smartregister.reveal.util.Utils.setTextViewText(facilityTextView, R.string.facility, facility);
-    }
-
-    @Override
-    public void setOperator() {
-        org.smartregister.reveal.util.Utils.setTextViewText(operatorTextView, R.string.operator, sharedPreferences.fetchRegisteredANM());
-    }
-
-    private void initializeProgressDialog() {
-        progressDialog = new ProgressDialog(this);
-        progressDialog.setCancelable(false);
-        progressDialog.setTitle(R.string.fetching_structures_title);
-        progressDialog.setMessage(getString(R.string.fetching_structures_message));
-    }
-
-    @Override
-    public void showProgressDialog(@StringRes int title, @StringRes int message) {
-        if (progressDialog != null) {
-            progressDialog.setTitle(title);
-            progressDialog.setMessage(getString(message));
-            progressDialog.show();
-        }
-    }
-
-    @Override
-    public void hideProgressDialog() {
-        if (progressDialog != null) {
-            progressDialog.dismiss();
-        }
-    }
 
     @Override
     public Location getUserCurrentLocation() {
@@ -634,6 +416,11 @@ public class ListTasksActivity extends BaseMapActivity implements ListTaskContra
         SyncStatusBroadcastReceiver.getInstance().removeSyncStatusListener(this);
         LocalBroadcastManager.getInstance(getApplicationContext()).unregisterReceiver(refreshGeowidgetReceiver);
         super.onPause();
+    }
+
+    @Override
+    protected BaseDrawerContract.Presenter getPresenter() {
+        return listTaskPresenter;
     }
 
     private class RefreshGeowidgetReceiver extends BroadcastReceiver {
